Only push gem if file containing version number is edited
This is to not spam my emails, but also make the action more efficient. The action fails if the version number is the same so if the file is edited in a way that is not a bump to the version number, the gem will not be pushed unless ready.