Remove unneeded div in navigation
The nav tag has the ID nav because I do not think there is another way to get the JavaScript working. This is because there is no function to grab a tag, and apply a style to it, in the same way that getElementById() does. Otherwise, the ID does literally nothing, as all the formatting is applied to the nav tag itself.